Use ipa and fix some EIP discrepancies

+++ empty.go
@@ -49,7 +49,7 @@ func (Empty) ComputeCommitment() *Fr {
 	return &FrZero
 }
 
-func (Empty) GetCommitmentsAlongPath([]byte) ([]*Point, []uint, []*Fr, [][]Fr) {
+func (Empty) GetCommitmentsAlongPath([]byte) ([]*Point, []uint8, []*Fr, [][]Fr) {
 	panic("trying to produce a commitment for an empty subtree")
 }
